README.md
Interactive list

HTML-document with an unordered list element, a text-box, and a button that says “Add.” JavaScript that appends a new element to the unordered list when the button is clicked. It has functionality that allow the to delete a selected list item. When the “delete” button is clicked, the item is removed from the list. When an item is added, add it to the local storage. When it is deleted; remove it. When the page loads, retrieve the items that exist in the offline storage to populate the list.